+++ src/main/java/com/carryjey/social/mapper/NotificationMapper.java
@@ -16,13 +16,13 @@ import java.util.Map;
 public interface NotificationMapper extends BaseMapper<Notification> {
 
     List<Map<String, Object>> selectByUserId(
-        @Param("userId") Integer userId, @Param("read") Boolean read, @Param("limit") Integer limit);
+        @Param("userId") long userId, @Param("read") Boolean read, @Param("limit") Integer limit);
 
     // 查询未读消息数量
     @Select("select count(1) from notification where target_user_id = #{userId} and `read` = false")
-    long countNotRead(@Param("userId") Integer userId);
+    long countNotRead(@Param("userId") long userId);
 
     // 将未读消息置为已读
     @Update("update notification set `read` = true where target_user_id = #{targetUserId}")
-    void updateNotificationStatus(@Param("targetUserId") Integer targetUserId);
+    void updateNotificationStatus(@Param("targetUserId") long targetUserId);
 }
