+++ src/main/java/com/carryjey/social/service/CodeService.java
@@ -43,7 +43,7 @@ public class CodeService {
     }
 
     // 查询没有用过的code
-    public Code selectNotUsedCode(Integer userId, String email) {
+    public Code selectNotUsedCode(long userId, String email) {
         QueryWrapper<Code> wrapper = new QueryWrapper<>();
         wrapper
             .lambda()
@@ -55,7 +55,7 @@ public class CodeService {
     }
 
     // 创建一条验证码记录
-    public Code createCode(Integer userId, String email) {
+    public Code createCode(long userId, String email) {
         Code code = this.selectNotUsedCode(userId, email);
         if (code == null) {
             code = new Code();
@@ -71,7 +71,7 @@ public class CodeService {
     }
 
     // 验证邮箱验证码
-    public Code validateCode(Integer userId, String email, String _code) {
+    public Code validateCode(long userId, String email, String _code) {
         QueryWrapper<Code> wrapper = new QueryWrapper<>();
         wrapper
             .lambda()
@@ -84,7 +84,7 @@ public class CodeService {
     }
 
     // 发送邮件
-    public boolean sendEmail(Integer userId, String email) {
+    public boolean sendEmail(long userId, String email) {
         Code code = this.createCode(userId, email);
         // 发送邮件
         return emailUtil.sendEmail(email, "修改邮箱验证码", "你的验证码是：" + code.getCode() + "<br>请在30分钟内使用");
